Generate grid and calculate container size
Using two "for" loops, one for width and one for height, the grid was
generated. Since flexbox is being used, using a wrap is enough to
have each div accomodate itself inside the container with wrap. After
that it is only necessary to calculate the width of the container by
multiplying the size of each square with the size of the grid