Adjust size of generated divs inside the grid
Each single "cell" that's being generated in the loop is adjusted
depending on the size of the grid to fit the container, this way, the
container will never be bigger or smaller than 480px but the squares
inside will change their size so they can fit an A x A sized grid.
This is done by dividing 480 by the grid size.