Add buttons to reset and change grid size
After confirming that each button can take an event listener,
a few functions were created to be able to make the necessary changes:

First, the code to generate the grid was put inside a function
expression, this way the grid can be generated again with a new size
from any part of the code.

Second, the reset function, which simply uses a selector to go
through all the divs and change the background of each one to white.

Third, the function to change the grid size which takes in a prompt
then it checks if the value that was entered is valid and doesn't
exceed the limit, if so, the current grid will be deleted and a new
one will be generated with the new size. Otherwise, the user will
receive a new prompt to enter a new valid number.