Create logic for random colors and color fade
By adding a boolean to check if the random colors are active, the code
from the grid generator uses a Math.random() multiplied by 255 inside
an rgb so each square of the grid will have a different color if
enabled

For the fade, opacity is changed using a value that increases by 0.09
until reaching 1, where it is capped by adding an if that checks the
value constantly. This value is also resetted if the random color
button is enabled or if the grid gets reset.